requirements.txt
kfp==2.7.0
kfp-pipeline-spec>=0.1.16
kfp-server-api>=1.8.5
kubernetes>=25.3.0
google-cloud-common>=1.2.0
google-cloud-notebooks>=1.7.1
google-cloud-aiplatform>=1.28.1
google-cloud-pipeline-components>=2.16.1
biopython>=1.81
google-cloud-filestore>=1.6.1
google-cloud-storage>=2.18.2